+++ test/test_data.c
@@ -3,11 +3,11 @@
 #include <assert.h>
 
 int main(int argc, char** argv) {
-    int num_entries = 0;
+    size_t num_entries = 0;
     char test_string[] = "hello=world\ngood=bye";
     struct Data* entries = y_data_parse(test_string, &num_entries);
 
-    for (int i = 0; i < num_entries; i++) {
+    for (size_t i = 0; i < num_entries; i++) {
         printf("%s is equal to %s\n", entries[i].name, entries[i].value);
     }
